requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
oauth2client==4.1.3
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
tornado==6.4
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
